pl - convert SummaryBox to TypeScript
