--- src/components/SummaryBox.d.ts
@@ -1,15 +1,0 @@
-import { CardGroupProps } from 'reactstrap/lib/CardGroup';
-import * as React from 'react';
-
-interface SummaryItem {
-  key?: string | number;
-  value: string;
-  label: string;
-}
-
-interface SummaryBoxProps extends CardGroupProps {
-  children?: ReactNode; 
-  items?: SummaryItem[];
-}
-declare class SummaryBox extends React.Component<SummaryBoxProps, {}> { }
-export default SummaryBox;
