+++ src/index.js
@@ -145,7 +145,7 @@ import Spinner from './components/Spinner.js';
 import StateInput from './components/StateInput.js';
 import StaticInput from './components/StaticInput.js';
 import Steps from './components/Steps';
-import SummaryBox from './components/SummaryBox.js';
+import SummaryBox from './components/SummaryBox.tsx';
 import SummaryBoxItem from './components/SummaryBoxItem.js';
 import Table from './components/Table.js';
 import TimeInput from './components/TimeInput.js';
